{"schema":"decodex/commit/1","summary":"Modularize live frame stream filter","authority":"manual"}
